scripts to help run tests and to set environment variables for path

+++ build-b-pc386.sh
@@ -5,16 +5,19 @@ TARGET=i386-rtems4.10
 EXTENSION=
 
 cd b-${BSP}${EXTENSION}
-rm * -rf
-#../rtems/configure --target=${TARGET} --disable-posix --disable-itron --disable-networking --disable-multiprocessing --enable-rtemsbsp=${BSP}
-#../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --disable-posix
-#../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --enable-tests=samples
-#../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --enable-tests
+if [[ $# -eq 0 ]]
+then
+  rm * -rf
+  #../rtems/configure --target=${TARGET} --disable-posix --disable-itron --disable-networking --disable-multiprocessing --enable-rtemsbsp=${BSP}
+  #../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --disable-posix
+  #../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --enable-tests=samples
+  #../rtems/configure --target=${TARGET} --disable-networking --enable-rtemsbsp=${BSP} --enable-tests
 
-## Build for QEMU
-../rtems/configure --target=${TARGET} --disable-networking \
-  --enable-rtemsbsp=${BSP} --enable-tests=samples \
-  USE_COM1_AS_CONSOLE=1 BSP_PRESS_KEY_FOR_RESET=0
+  ## Build for QEMU
+  ../rtems/configure --target=${TARGET} --disable-networking \
+    --enable-rtemsbsp=${BSP} --enable-tests=samples \
+    USE_COM1_AS_CONSOLE=1 BSP_PRESS_KEY_FOR_RESET=0
+fi
 
 time make -j 4 2> make.error
